requirements.txt
# Core dependencies
streamlit==1.32.0
anthropic==0.8.1
pandas==2.2.0
numpy==1.26.3

# Data visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Data generation and manipulation
faker==22.5.0
scipy==1.12.0

# Environment and configuration
python-dotenv==1.0.1
python-decouple==3.8

# Testing
pytest==8.0.0
pytest-cov==4.1.0

# Type checking and code quality
mypy==1.8.0
black==24.1.1
flake8==7.0.0
isort==5.13.2

# Data handling and processing
openpyxl==3.1.2  # For Excel support in pandas
pyarrow==15.0.0  # For better performance with pandas

# Date and time handling
pytz==2024.1
python-dateutil==2.8.2

# Progress bars and spinners
tqdm==4.66.1
requests==2.31.0

# JSON handling
ujson==5.9.0

# Additional Streamlit components
streamlit-option-menu==0.3.6
streamlit-metrics==0.1.0
streamlit-aggrid==0.3.4
plotly-resampler==0.9.1

# System and path handling
pathlib==1.0.1

# Color handling for visualizations
colorama==0.4.6
colour==0.1.5

# Performance optimization
numba==0.58.1